% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico.R
\name{git}
\alias{git}
\title{Automatizar Renderizacao e Comandos do Git}
\usage{
git(comentario)
}
\arguments{
\item{comentario}{mensagem do commit. Se omitido, envia um espaco em branco.}
}
\value{
Executa os comandos no sistema.
}
\description{
\code{git} executa o quarto render e sequencialmente os comandos add, commit e push.
}
\examples{
library(graf)
# git()
# git(versao_3)
}
