% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico.R
\name{mdes}
\alias{mdes}
\title{Cálculo de Desbalanceamento de Correntes}
\usage{
mdes(correntes)
}
\arguments{
\item{correntes}{Vetor numérico com as correntes (ex: c(13.5, 12.1, 14.0)).}
}
\value{
O desbalanceamento em porcentagem.
}
\description{
\code{mdes} calcula o desbalanceamento percentual baseado na razão entre a maior e a menor corrente lida, alertando se ultrapassar 5%.
}
\examples{
library(graf)
# mdes(c(14.8, 15.2, 14.1))
}
