% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico.R
\name{mpot}
\alias{mpot}
\title{Cálculo de Potência de Motores Elétricos}
\usage{
mpot(fases, v, i, fp, rendimento = NULL)
}
\arguments{
\item{fases}{Número de fases (1, 2 ou 3).}

\item{v}{Tensão de alimentação (Volts).}

\item{i}{Corrente elétrica (Amperes).}

\item{fp}{Fator de potência.}

\item{rendimento}{Rendimento do motor (opcional, entre 0 e 1).}
}
\value{
Lista com o tipo de potência e valores em kW e CV.
}
\description{
\code{mpot} calcula a potência elétrica ativa ou a potência mecânica no eixo (se o rendimento for informado).
}
\examples{
library(graf)
# Potência elétrica ativa (sem rendimento) para motor trifásico
# mpot(3, 380, 14.8, 0.85)

# Potência mecânica no eixo (com rendimento)
# mpot(3, 380, 14.8, 0.85, 0.90)
}
