\name{print.graf}
\alias{print.graf}

\title{
Print a \code{graf} object.
}

\description{
Prints a brief summary of a \code{graf} object.
}
\usage{
\method{print}{graf}(x, ...)
}

\arguments{
  \item{x}{
A \code{graf} object.
}
  \item{\dots}{
Additional arguments to pass to \code{print}
}
}

\value{
The number of presence and absence/background records and covariates used to fit the model.
}

\seealso{
\code{\link{graf}}
}